requirements.txt
braintrust==0.2.6
openai>=1.12.0
anthropic>=0.7.0
google-generativeai>=0.3.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
jinja2>=3.1.2
httpx>=0.25.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
